ihaCore/requirements.txt
Django==5.0.4
django-datatable==2.1.0
djangorestframework==3.15.1
psycopg2==2.9.9